rename and clean up code
